Factor out new LoaderPool class and file
